pyproject.toml
name = "atomicio"
version = "0.1.0"
description = "Atomic locked file operations with plugin format support"
authors = ["Santiago Matta <santiagomatta@outlook.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
filelock = ">=3.12.0"
pyyaml = ">=6.0.1"
tomli = {version = ">=2.0.1", python = "<3.11"}
tomli_w = {version = ">=1.0.0", python = "<3.11"}
atomicwrites = ">=1.4.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"

[tool.poetry.urls]
homepage = "https://github.com/santiago897/atomicio"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.classifiers]
classifiers = [
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.8",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
	"Development Status :: 4 - Beta",
	"Intended Audience :: Developers",
	"Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.extras]
yaml = ["pyyaml"]
toml = ["tomli", "tomli_w"]

[tool.poetry.scripts]
atomicio = "atomicio.__main__:main"

[tool.poetry.entry-points."atomicio.formats"]
# Ejemplo: otros paquetes pueden registrar aquí sus formatos
# myplugin = "myplugin.formats:register"